refactor(models): DiagnosticExtract renames and DiagnosticMapping.ServiceMapping per spec (Rule 0007 A)
