+++ src/armodel/models/M2/AUTOSARTemplates/BswModuleTemplate/BswBehavior.py
@@ -28,7 +28,7 @@ from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.PrimitiveTypes import RefType
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.Identifiable import Referrable
 from armodel.models.M2.AUTOSARTemplates.SWComponentTemplate.Datatype.DataPrototypes import VariableDataPrototype
-from armodel.models.M2.AUTOSARTemplates.SWComponentTemplate.RPTScenario import IdentCaption
+from armodel.models.M2.AUTOSARTemplates.DiagnosticExtract.DiagnosticMapping.ServiceMapping import BswServiceDependencyIdent
 from armodel.models.M2.AUTOSARTemplates.SWComponentTemplate.SwcInternalBehavior.IncludedDataTypes import IncludedDataTypeSet
 from armodel.models.M2.AUTOSARTemplates.SWComponentTemplate.SwcInternalBehavior.ModeDeclarationGroup import IncludedModeDeclarationGroupSet
 from armodel.models.M2.AUTOSARTemplates.CommonStructure.ModeDeclaration import ModeActivationKind
@@ -2107,28 +2107,6 @@ class BswInternalTriggeringPoint(Identifiable):
         if value is not None:
             self.swImplPolicy = value
         return self
-
-
-class BswServiceDependencyIdent(IdentCaption):
-    """
-    This meta-class is created to add the ability to become the target of a reference
-    to the non-Referrable BswServiceDependency.
-    """
-
-    # BswServiceDependencyIdent method parity checklist:
-    # Spec: AUTOSAR_CP_TPS_DiagnosticExtractTemplate.pdf, Table 5.16, p.240
-    # Spec verified: R23-11
-    # [x] __init__                     [x] impl  [x] docstring  [x] test
-
-    def __init__(self, parent: ARObject, short_name: str):
-        """
-        Initializes the BswServiceDependencyIdent with a parent and short name.
-
-        Args:
-            parent: The parent ARObject that contains this ident caption
-            short_name: The unique short name of this ident caption
-        """
-        super().__init__(parent, short_name)
 
 
 class RoleBasedBswModuleEntryAssignment(ARObject):
