+++ src/armodel/models/M2/AUTOSARTemplates/DiagnosticExtract/DiagnosticContribution.py
@@ -12,7 +12,7 @@ from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.ArObject import (
     ARObject,
 )
-from armodel.models.M2.AUTOSARTemplates.DiagnosticExtract.DiagnosticCommonElement import DiagnosticCommonElement
+from armodel.models.M2.AUTOSARTemplates.DiagnosticExtract.CommonDiagnostics import DiagnosticCommonElement
 
 
 class DiagnosticServiceTable(DiagnosticCommonElement):
