+++ src/armodel/models/__init__.py
@@ -166,7 +166,8 @@ from armodel.models.M2.AUTOSARTemplates.CommonStructure.Timing.TimingDescription
 
 # Additional DiagnosticExtract imports
 from armodel.models.M2.AUTOSARTemplates.DiagnosticExtract import *  # noqa: F403
-from armodel.models.M2.AUTOSARTemplates.DiagnosticExtract.DiagnosticCommonElement import *  # noqa: F403
+from armodel.models.M2.AUTOSARTemplates.DiagnosticExtract.CommonDiagnostics import *  # noqa: F403
+from armodel.models.M2.AUTOSARTemplates.DiagnosticExtract.DiagnosticMapping.ServiceMapping import *  # noqa: F403
 
 # ECUCParameterDefTemplate
 from armodel.models.M2.AUTOSARTemplates.ECUCParameterDefTemplate import *  # noqa: F403
